+++ src/app/features/cart/cart.module.ts
@@ -1,4 +1,4 @@
-import { NgModule } from '@angular/core';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
@@ -8,11 +8,14 @@ import { SharedModule } from 'src/app/shared/shared.module';
 const EXPORTED_COMPONENTS = [
   CartListComponent
 ];
+const LOCAL_COMPONENTS = [
+  CartItemComponent
+];
 
 @NgModule({
   declarations: [
     ...EXPORTED_COMPONENTS,
-    CartItemComponent
+    ...LOCAL_COMPONENTS
   ],
   imports: [
     CommonModule,
@@ -21,6 +24,7 @@ const EXPORTED_COMPONENTS = [
   ],
   exports: [
     ...EXPORTED_COMPONENTS
-  ]
+  ],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class CartModule { }