+++ src/app/shared/directives/no-scroll.directive.ts
@@ -1,0 +1,12 @@
+import { Directive, HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[shpNoScroll]'
+})
+export class NoScrollDirective {
+  @HostListener('wheel', ['$event'])
+  onWheel(event: Event) {
+    event.preventDefault();
+  }
+
+}