+++ src/app/shared/shared.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './directives/highlight.directive';
+import { NoScrollDirective } from './directives/no-scroll.directive';
 
 const EXPORTED_COMPONENTS = [
   HighlightDirective,
+  NoScrollDirective
 ];
 
 @NgModule({