+++ changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2021-09-27
+### Removed
+### Changed
+- Renamed some fields of CartService
+- SetQuantity to decreaseAmount / increaseAmount methods
+### Added
+- CartService: increaseQuantity, decreaseQuantity, isEmptyCart, removeAllProducts
+- NoScrollDirective
+
 ## [0.2.0] - 2021-09-15
 ### Removed
 - Test First component